backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
supabase
asyncpg
pymongo
python-dotenv
requests
email-validator